requirements.txt
streamlit
pdf2image
# sudo apt-get update
# sudo apt-get install poppler-utils

pillow
langchain
langchain_community
openai

